Multiple commits in Super Quick Patch Guide
How to use perlbug when a change is a series of commits, not a single
commit.

This is the advice RJBS gave me over IRC. Including it in the guide should
avoid him having to repeat the advice to others.

Committer: Added single quotes around one keyboard command.

For: RT #119599