Intercept assignment to $! to translate WSAExxx values to Exxx values on Windows
Since perl previously assigned WSAExxx values to $! on Windows it is quite
possible that (Perl-level) user code may also manually make similar assignments,
which will now cause breakage if the value put in $! is subsequently compared to
Errno/POSIX constants because the latter are now the corresponding Exxx values
where possible.

An example of this is in Net::Ping::tcp_connect(), which does the following to
fetch a socket-level error code:

unpack("i", getsockopt($self->{"fh"}, SOL_SOCKET, SO_ERROR))

and assigns the result (a WSAExxx value such as 10061) to $! and then goes wrong
in the subsequent test (in ping_tcp()) for $! == ECONNREFUSED (which is now 107
rather than 10061 if perl is built with VC10 or higher).

To avoid this we now intercept assignment to $! and convert any WSAExxx values
to Exxx values first.  This causes a minor oddity in that this:

perl -le "$! = 10061; print 0+$!"

will now output 107 (for VC10+ perls) but this is surely preferable to the
alternative breakage described above.