Simplify errno2.h slightly
There is no need for it to worry about ensuring that standard errno.h values
like EINTR are defined.

There is no point in it defining dummy constants like EINVALIDPROCTABLE which
do not exist in errno.h at all. They would only be given the corresponding
WSAExxx values anyway, so the win32sck.c function might just as well return
them directly.

Other constants like ESOCKTNOSUPPORT which do not exist in errno.h on Windows
but presumably exist elsewhere since POSIX tries to exports them are still
worth defining (where possible -- it isn't possible for EHOSTDOWN since there
is no corresponding WSAEHOSTDOWN); add comments for these for clarity.

Also comment which errno.h constants are new in VC10, and mention three of
the four that win32/include/sys/socket.h used to redefine because they are
used in the perl core (ENOTSOCK, EAFNOSUPPORT and ECONNABORTED) -- the other
one (ECONNRESET) does not appear to be used in the perl core as far as I can
tell.