[perl #79074] make sure HeSVKEY_force() preserves UTF8ness (pvn)
HeSVKEY_force() is only used in two places in core.

In the first case, the key is always stored as a SV (when handling tie
magic, since NEXTKEY can only return a SV)

The second case is in B::HE, but I don't see a way to create a B::HE object
from a hash.