Remove the hints for Kubota Pacific's Titan 3000 Series Machines.
These appear to be a hold-over from Perl 4 days. We think it unlikely that
perl 5.000 ever compiled correctly on these machines, and changes in 5.004
would definitely have made Perl unbuildable. Given that we've had no bug
reports in 16 years, it's clear that this platform is completely unused,
and thus safe to remove it without any warning period.

See RT #119745 for more details.