perldelta: Two to-dos that don’t deserve mention
9cefd268 [perl #118747] Allow in-place s///g when !!PL_sawampersand
c04e4151 op.c:leave_scope: use mg_free before sv_force_normal [#119295]

In both cases, it is a module relying on undocumented internals.

In the former, we do have an optimisation kicking in in cases where it
didn’t in 5.18 and earlier.  But those cases are rare enough that an
entry in ‘Performance Enhancements’ would be misleading.

In the latter, the order in which clearing and ‘free’ magic are trig-
gered is more predictable, but I don’t believe that has ever been
guaranteed, and only creepy modules will be affected by it.