toke.c, scan_ident: Changes to whitespace skipping
Rather than conditionally calling PEEKSPACE(), always call
skipspace_flags, but pass LEX_NO_SWALLOW_COMMENTS on the first use,
when skipping over whitespace after the open-bracket in ${...},
so that ${#} is not misparsed.

Additionally, this commit gives scan_ident() a flags parameter;
for now, those flags are passed directly to skipspace_flags().
This is used by intuit_more() to signal that, even if we encounter
the end of the buffer, we shouldn't try grabbing the next chunk,
since intuit_more() already has that and gave scan_ident() only
as much as it needed.

This fixes [perl #119889].