perldoc.pod should be copied to lib/perldoc.pod, not lib/Pod/perldoc.pod
MakeMaker's default file scanning rules will cause it to be staged to lib/Pod
which would cause it to be installed to lib/Pod/perldoc.pod. It *isn't*
installed there by core perl, as the "duplicate Pod" detection code spots
that it has already been installed to .../pod/ along with the Pods from pod/,
but the CPAN distribution does install it to a place that perldoc (the
program) doesn't search.

It's surprising that no-one has noticed this before.