corelist.pl - Fix the addition of a new perl release following 9f92b9bec5
Now that the check is working, it was working slightly too well and actually
picked up 5.019005 in the %delta because that had been added already and
thus didn't add it to %released when it should have done!

One simple fix is to move the processing of %released so that it's done
first, which makes sense since it appears first in the file anyway.