PATCH: [perl #120041] regcomp.c missing parens and broken STCLASS
This was caused by a problem in commit
43a64b8b430604bd4d76fd256a5babdccaf0ab2b
which was masked by problems in commit
a0dd42312a1f26356d2fdf49656e45b77c2cefb5.

The first commit deleted a line it shouldn't have.  I thought that a
variable wasn't modified before reaching that line, but it turns out it
was because another variable is set to point to its address, and was
modified inside a subroutine call.  The other two lines deleted in the
first commit are ok to have deleted.

The second commit masked the problems of the first by omitting some
necessary grouping parentheses, which caused things to not work as
intended, and we didn't have a good test case in our suite to find this
problem.  (Thanks to Lukas Mai for spotting the problem and submitting a
test case.)