podcheck.t: Make blead pass again
Blead is failing sometimes on podcheck.t.  I don't fully understand the
reasons, but it is some sort of race condition.

The base problem that is causing this race to show up is a bug in the
new version 3.29 of Test::Harness.  It includes 2 pods that have the
same name.  This creates problems for reading the documentation in an
html browser.  The software that creates the hyperlinks uses the pod
name to figure out what the linked-to document is.  It cannot figure out
which one to link to when multiple pods have the same name, as here, so
podcheck raises an error.  Test::Harness should change so that the pods
have different names.  It's like having two different Unix commands
named 'foobar'.  Which one should `man foobar` display, if it could only
display one?

I haven't figured out how to fix the race yet, so this commit just
ignores one of the conflicting files for now.  The race has something to
do with the fact that these are both cpan files which are not normally
examined for pod problems, and so the error for one of them is getting
suppressed; and which one of the two gets suppressed varies.

This bug in Test::Harness also manifested itself as showing up a broken
link.