Porting/Maintainers.pl - Change UPSTREAM => undef to 'blead' for dist/ modules
There is only one module affected (XSLoader), already having a note to
revert it to 'blead' once 0.17 is released. That version is not yet
released, but blead is identical to 0.16, and I believe that note was not
quite correct -- it was a new *perl* release that was needed, not a new
CPAN release of XSLoader. The comment for the commit (e52529ca24) which
added the note is correct: "The problem (cmp_version.t failing) should
disappear after the next perl release." That was back when 5.17.3 had just
been released, and the problem has indeed gone away now.