$^E should have same handling as $! for Win32 and OS/2
Commit 1500bd919ffeae0f3252f8d1bb28b03b043d328e changed the handling of
$! to look for UTF-8 messages.  This issue is also present in Win32 for
$^E.  The two should have uniform treatment, so this commit causes the
actual same code to be executed for both.  OS/2 is also subject to
locale issues, and so it also is changed here to use the same code, so
that future changes will apply to it automatically.

VMS doesn't use locales, so it retains its current behavior.

Note that 1500bd919 has created some field problems, so that the changes
it introduced will be further changed or reverted.  The current commit
just makes sure that whatever those further changes are will be
automatically propagated to all necessary platforms.